feat(protocol): widen item ids to u32 across trade-related parses
The server sends item ids with addItemId (u32 under GameU32ItemIds), so the
matching reads must use getItemId / writes addItemId instead of the hardcoded
U16, otherwise ids above 65535 truncate. Also widen MarketData.tradeAs/showAs
to u32 (the protobuf path must not down-cast; the legacy 8.60 .dat path stays
U16, byte-identical).

- cyclopedia: bestiary loot, item summary, concoctions
- dailyreward: pick-items list and claim columns
- forge: fusion/transfer/donors/receivers and result ids
- stash: withdraw/stow requests and open-stash parse
- market: enter/detail/own-offer/browse/create
- appearancesloader + thingtype: MarketData tradeAs/showAs as u32
- protocolgameparse: proficiency screenshot item id

Target: OTC

+++ src/client/appearancesloader.cpp
@@ -422,10 +422,10 @@ void AppearancesLoader::applyFlags(const AppearanceFlags& f, const std::string& 
         md.name           = name;
         md.category       = m.has_category() ? static_cast<int>(m.category()) : 0;
         md.tradeAs        = m.has_trade_as_object_id()
-                              ? static_cast<uint16>(m.trade_as_object_id())
+                              ? m.trade_as_object_id() // u32: item ids > 65535 must not truncate
                               : 0;
         md.showAs         = m.has_show_as_object_id()
-                              ? static_cast<uint16>(m.show_as_object_id())
+                              ? m.show_as_object_id()
                               : 0;
         // Legacy stores a single restrictVocation U16. Proto sends a repeated
         // list; keep the legacy collapse for the .dat path AND the full list for
